pyproject.toml
[project]
name = "rag-zotero"
version = "0.1.0"
description = "Extract text from Zotero storage and enable RAG-style local search."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
  "pymupdf>=1.23",
  "chromadb>=0.5.0",
  "openai>=1.30.0",
]

[project.optional-dependencies]
local-embeddings = [
  "sentence-transformers>=2.6.0",
]

[project.scripts]
rag-zotero = "rag_zotero.cli:app"

[build-system]
requires = ["hatchling>=1.22"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_zotero"]

[tool.ruff]
line-length = 100

